+++ datapkg/index/__init__.py
@@ -2,6 +2,13 @@
 filesystem, database, CKAN).
 '''
 from base import SimpleIndex, FileIndex
-from db import DbIndex, Index
+from db import DbIndexSqlite, DbIndexSqlalchemy
 from ckan import CkanIndex
+
+# TODO: would be nice for this to be a property named default_index
+def get_default_index():
+    idx = DbIndexSqlite()
+    idx.init()
+    return idx
+